correction du type de donnees

+++ chapitre-01/double-doubler_un_entier/main.cpp
@@ -1,10 +1,10 @@
 #include <iostream>
 
 int main(){
-    float a,b;
+    int a,b;
     std::cout <<"entrer un nombre"<<std::endl;
     std::cin>> a;
-    float d=a*2;
+    int d=a*2;
     std::cout << d<<std::endl;
     return 0;
 }
